% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dburr.R
\name{dburr}
\alias{dburr}
\title{Burr Type XII Density Function}
\usage{
dburr(x, c, k)
}
\arguments{
\item{x}{Numeric vector of values (x > 0).}

\item{c}{Positive shape parameter (c > 0).}

\item{k}{Positive shape parameter (k > 0).}
}
\value{
Numeric vector of density values f(x).
}
\description{
Computes the probability density function (PDF) of the Burr Type XII distribution.
}
\details{
The density function is defined as:
\deqn{f(x) = \frac{c k x^{c-1}}{(1 + x^c)^{k+1}}, \quad x > 0}
}
\examples{
dburr(1, c = 2, k = 3)

}
