% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pburr.R
\name{pburr}
\alias{pburr}
\title{Burr Type XII Cumulative Distribution Function}
\usage{
pburr(x, c, k)
}
\arguments{
\item{x}{Numeric vector of values (x > 0).}

\item{c}{Positive shape parameter (c > 0).}

\item{k}{Positive shape parameter (k > 0).}
}
\value{
Numeric vector of cumulative probabilities F(x).
}
\description{
Computes the cumulative distribution function (CDF) of the Burr Type XII distribution.
}
\details{
The cumulative distribution function is defined as:
\deqn{F(x) = 1 - (1 + x^c)^{-k}, \quad x > 0}
}
\examples{
pburr(1, c = 2, k = 3)

}
