% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rburr.R
\name{rburr}
\alias{rburr}
\title{Random Generation from Burr Type XII Distribution}
\usage{
rburr(n, c, k)
}
\arguments{
\item{n}{Integer. Number of observations to generate.}

\item{c}{Positive shape parameter (c > 0).}

\item{k}{Positive shape parameter (k > 0).}
}
\value{
Numeric vector of random values distributed as Burr(c, k).
}
\description{
Generates random values from the Burr Type XII distribution using
the inverse transform method.
}
\details{
Random values are generated using the inverse transform method:
\deqn{X = \left((1 - U)^{-1/k} - 1\right)^{1/c}, \quad U \sim U(0,1)}
}
\examples{
x <- rburr(1000, c = 2, k = 3)
hist(x, probability = TRUE)
curve(dburr(x, 2, 3), add = TRUE, col = "red")

}
